Renamed FileUploadBehavior to AttachmentBlacklistBehavior
Moved to Modix.Services.Moderation
Made handle method private
